+++ NEWS.md
@@ -1,3 +1,10 @@
+# peskas.kenya.data.pipeline 0.3.0
+
+## New features
+
+- Add ingestion function to ingest ongoing data `ingest_surveys()`
+- Add preprocessing function to preprocess ongoing data `preprocess_landings()`
+
 # peskas.kenya.data.pipeline 0.2.0
 
 - Add package website with documentation functions